requirements.txt
pydantic-ai
pydantic-ai-slim[duckduckgo]
pytest
pytest-dotenv